Ensure that ratio_col is not indexed with a negative value (#112)
I'm not sure what exactly caused it but my installation of rTorrent-PS was consistently segfaulting here. I fiddled with it a bit and it seemed that it was indexing with a negative value. As far as I can tell, there's no way for `rcol` to be negative so it must have been the ratio (couldn't tell for sure because `ratio` was optimized out and not viewable in the debugger).