Use distro automake's config.guess
Closes https://github.com/pyroscope/rtorrent-ps/issues/137

The config.guess that ships with the version of xmlrpc-c we use is quite
outdated - it's from 2009. It can't identify some modern systems,
specifically ARM64 (i.e aarch64. x86_64 is obviously okay, I expect armv7l
also as there is some mention of RPi 3's in the script).
To work around this without using a different version of xmlrpc-c, we can
simply copy in the newer config.guess that ships with the version of
automake we install from the distro. This results in a successful build
on newer platforms.